PWA has 100% score now in prod env
