+++ package.json
@@ -8,8 +8,8 @@
   "license": "ISC",
   "private": false,
   "scripts": {
-    "dev": "webpack-dev-server --color",
-    "build": "webpack -p",
+    "dev": "webpack-dev-server --color --open --config webpack.dev.js",
+    "build": "webpack -p --config webpack.prod.js",
     "test": "karma start"
   },
   "dependencies": {
@@ -29,6 +29,7 @@
     "babel-loader": "^7.1.2",
     "babel-preset-env": "^1.6.1",
     "babel-preset-react": "^6.24.1",
+    "clean-webpack-plugin": "^0.1.17",
     "copy-webpack-plugin": "^4.1.1",
     "css-loader": "^0.28.7",
     "extract-text-webpack-plugin": "^3.0.1",
@@ -50,8 +51,10 @@
     "redux-devtools-extension": "^2.13.2",
     "redux-logger": "^3.0.6",
     "style-loader": "^0.19.0",
+    "uglify-js-plugin": "^0.0.6",
     "url-loader": "^0.6.2",
     "webpack": "^3.8.1",
-    "webpack-dev-server": "^2.9.2"
+    "webpack-dev-server": "^2.9.2",
+    "webpack-merge": "^4.1.0"
   }
 }
