--- src/config/history.js
@@ -1,3 +1,0 @@
-import createHistory from 'history/createBrowserHistory'
-
-export const history = createHistory()
