+++ src/config/manifest.json
@@ -1,10 +1,20 @@
 {
-  "short_name": "DFYG",
-  "name": "DFYG PWA",
+  "short_name": "DFYG PWA",
+  "name": "DFYG Progressive Web App",
   "icons": [
     {
-      "src": "favicon.ico",
+      "src": "/images/splash_icon_192.png",
       "sizes": "192x192",
+      "type": "image/png"
+    },
+    {
+      "src": "/images/splash_icon_256.png",
+      "sizes": "256x256",
+      "type": "image/png"
+    },
+    {
+      "src": "/images/splash_icon_512.png",
+      "sizes": "512x512",
       "type": "image/png"
     }
   ],
