+++ src/data/configureStore.js
@@ -2,9 +2,8 @@ import { routerMiddleware } from 'react-router-redux'
 import { createStore, applyMiddleware } from 'redux'
 import { composeWithDevTools } from 'redux-devtools-extension'
 import { createLogger } from 'redux-logger'
-import { createEpicMiddleware } from 'redux-observable';
+import { createEpicMiddleware } from 'redux-observable'
 
-import { history } from '../config/history'
 import rootReducer from './reducers/rootReducer'
 import rootEpic from './actions/rootEpics'
