+++ webpack.prod.js
@@ -1,0 +1,21 @@
+const webpack= require('webpack')
+const merge = require('webpack-merge')
+const CleanWebpackPlugin = require('clean-webpack-plugin')
+const UglifyJSPlugin = require('uglifyjs-webpack-plugin')
+
+const common = require('./webpack.common.js')
+
+module.exports = merge(common, {
+  devtool: 'nosources-source-map',
+  plugins: [
+    new CleanWebpackPlugin(['dist']),
+
+    new UglifyJSPlugin({
+      sourceMap: true
+    }),
+
+    new webpack.DefinePlugin({
+     'process.env.NODE_ENV': JSON.stringify('production')
+    })
+  ]
+})
